Added the _phpi_ fence_custom to remarkable for php examples
Our core class documentation is littered with little code snippets and a
corresponding example, an SVG image that shows it's effect.

We serve these ups in tabs, but writing the HTML to construct them
inside markdown is a pita and also a bit much to ask for a non-
technical editor.

In addition, standard bootstrap tabs depend on jQuery, but that's
not (reliably) available in React.

Remarkable allows for custom fence blocks of code and that's what we
use here to implement out examples.
Use:

```phpx someidhere
example code here
```

And two tabs will be generated, the first will hold someidhere.svg
that must be present in the same folder as the markdown file.
The second will have the highlighted code.