First feeble attempts at getting redux to do something useful
If you open the console, you'll notice that clicking the
menu icon (the hamburger) triggers the TOGGLE_MENU redux
action, and that the _shown_ state is toggled between
true and false.

So far so good, but this state is not passed to the MainMenu
component, which means that it doesn't actually do anything yet.